policy, cmd: Add list-principals sub command
