+++ docs/cli/gittuf_policy_list-principals.md
@@ -1,0 +1,30 @@
+## gittuf policy list-principals
+
+List principals for the current policy in the specified rule file
+
+```
+gittuf policy list-principals [flags]
+```
+
+### Options
+
+```
+  -h, --help                 help for list-principals
+      --policy-name string   specify rule file to list principals for (default "targets")
+      --policy-ref string    specify which policy ref should be inspected (default "policy")
+```
+
+### Options inherited from parent commands
+
+```
+      --profile                      enable CPU and memory profiling
+      --profile-CPU-file string      file to store CPU profile (default "cpu.prof")
+      --profile-memory-file string   file to store memory profile (default "memory.prof")
+  -k, --signing-key string           signing key to use to sign policy file
+      --verbose                      enable verbose logging
+```
+
+### SEE ALSO
+
+* [gittuf policy](gittuf_policy.md)	 - Tools to manage gittuf policies
+
